docs(ingest): refactor docgen process (#12300)
