--- metadata-ingestion/docs/sources/powerbi/powerbi-report-server_pre.md
@@ -1,13 +1,0 @@
-## Configuration Notes
-See the 
-1. [Microsoft Grant user access to a Report Server doc](https://docs.microsoft.com/en-us/sql/reporting-services/security/grant-user-access-to-a-report-server?view=sql-server-ver16)
-2. Use your user credentials from previous step in yaml file
-## Concept mapping 
-
-| Power BI Report Server    | Datahub             |                                                                                               
-| ------------------------- | ------------------- |
-| `Paginated Report`        | `Dashboard`         |
-| `Power BI Report`         | `Dashboard`         |
-| `Mobile Report`           | `Dashboard`         |
-| `Linked Report`           | `Dashboard`         |
-| `Dataset, Datasource`     | `N/A`               |
